requirements.txt
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
Django==2.2.4
django-bootstrap-form==3.4
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-timezone-field==3.1
djangorestframework==3.10.3
html5lib==1.0.1
importlib-metadata==0.23
kombu==4.6.5
more-itertools==7.2.0
Pillow==6.2.0
PyPDF2==1.26.0
python-crontab==2.3.9
python-dateutil==2.8.0
pytz==2019.2
redis==3.3.11
reportlab==3.5.28
six==1.12.0
sqlparse==0.3.0
vine==1.3.0
webencodings==0.5.1
xhtml2pdf==0.2.3
xlwt==1.3.0
zipp==0.6.0
django-extensions==2.1.6
django-widget-tweaks==1.4.3
python-decouple==3.1